+++ shared/src/main/java/io/github/idankoblik/jukebox/Playlist.java
@@ -12,12 +12,18 @@ public class Playlist<P extends Platform> {
     private final Queue<AbstractSong<P>> songs = new LinkedList<>();
     private boolean playing = false;
 
-    
+    /**
+     * Adds a song to the playlist.
+     *
+     * @param song the song to be added to the playlist.
+     */
     public void addSong(AbstractSong<P> song) {
         songs.offer(song);
     }
 
-    
+    /**
+     * Starts playing the songs in the playlist. If no songs are available or a song is already playing,
+     */
     public void playSongs() {
         if (songs.isEmpty() || playing)
             return;
@@ -26,7 +32,9 @@ public class Playlist<P extends Platform> {
         playNextSong();
     }
 
-    
+    /**
+     * Plays the next song in the playlist. If there are no songs left, the playlist stops playing.
+     */
     private void playNextSong() {
         if (!playing || songs.isEmpty()) {
             playing = false;
@@ -43,24 +51,36 @@ public class Playlist<P extends Platform> {
         });
     }
 
-    
+    /**
+     * Stops all songs in the playlist and clears the playlist.
+     */
     public void stopSongs() {
         playing = false;
         songs.forEach(AbstractSong::stopSong);
         songs.clear();
     }
 
-    
+    /**
+     * Returns whether any song is currently being played.
+     *
+     * @return true if a song is playing, false otherwise.
+     */
     public boolean isPlaying() {
         return playing;
     }
 
-    
+    /**
+     * Returns the number of remaining songs in the playlist.
+     *
+     * @return the number of remaining songs in the playlist.
+     */
     public int getRemainingSongs() {
         return songs.size();
     }
 
-    
+    /**
+     * Clears the playlist by removing all songs.
+     */
     public void drop() {
         this.songs.clear();
     }